chore: swap the const value of server hello reply
by setting StatusTCPOnly == 1, when a new version client connect to an
old version server, the client will treat the server as udp disabled and
will not submit a udp request or open a udp control stream.

this make the upgrade much more smoothly even if we don't change the
protocolVersion.